+++ handlers/ajax_unit_handler.php
@@ -4,26 +4,28 @@ require_once('../../../../config.php');
 global $DB;
 
 if (isset($_POST)) {
-    $units = $_POST["units"];
-    $currentunits = \format_udehauthoring\model\unit_config::instance_all();
-    $savedunits = [];
-    foreach ($units as $key => $unitvalue) {
-        if($key > 0) {
-            $unit = \format_udehauthoring\model\unit_config::instance_by_name('udeh_unit_' . $key);
-            if ($unit === null) {
-                $unit = new \format_udehauthoring\model\unit_config();
-                $unit->name = 'udeh_unit_' . $key;
+    if (array_key_exists("units", $_POST)) {
+        $units = $_POST["units"];
+        $currentunits = \format_udehauthoring\model\unit_config::instance_all();
+        $savedunits = [];
+        foreach ($units as $key => $unitvalue) {
+            if($key > 0) {
+                $unit = \format_udehauthoring\model\unit_config::instance_by_name('udeh_unit_' . $key);
+                if ($unit === null) {
+                    $unit = new \format_udehauthoring\model\unit_config();
+                    $unit->name = 'udeh_unit_' . $key;
+                }
+                $unit->value = $unitvalue;
+                $unit->save();
             }
-            $unit->value = $unitvalue;
-            $unit->save();
+            $savedunits[] = $unitvalue;
         }
-        $savedunits[] = $unitvalue;
-    }
 
-    if(count($currentunits) > 0) {
-        if(count($savedunits) < count($currentunits)) {
-            for($i = (count($savedunits) - 1); $i < (count($currentunits) - 1); $i++) {
-                $currentunits[$i]->delete();
+        if(count($currentunits) > 0) {
+            if(count($savedunits) < count($currentunits)) {
+                for($i = (count($savedunits) - 1); $i < (count($currentunits) - 1); $i++) {
+                    $currentunits[$i]->delete();
+                }
             }
         }
     }
